xfree86: os-support: bsd: move wscons open logic into console_wscons.c
Yet another step on factoring out console type specific code so
separate drivers.
