+++ hw/xfree86/os-support/bsd/xf86_bsd_priv.h
@@ -5,6 +5,8 @@
 #ifndef _XSERVER_XFREE86_OS_SUPPORT_BSD_PRIV_H
 #define _XSERVER_XFREE86_OS_SUPPORT_BSD_PRIV_H
 
+#include <stdbool.h>
+
 #ifdef __OpenBSD__
 #define DEV_MEM "/dev/xf86"
 #else
@@ -23,6 +25,7 @@ void xf86_bsd_acquire_vt(void);
 
 /* PCVT console driver */
 void xf86_console_pcvt_close(void);
+bool xf86_console_pcvt_open(void);
 void xf86_console_pcvt_reactivate(void);
 
 /* SYSCONS console driver */
@@ -31,6 +34,7 @@ void xf86_console_syscons_reactivate(void);
 
 /* WSCONS console driver */
 void xf86_console_wscons_close(void);
+bool xf86_console_wscons_open(void);
 
 extern int initialVT;
